+++ MoltenVK/MoltenVK/API/vk_mvk_moltenvk.h
@@ -780,24 +780,24 @@ typedef struct {
 	 * Controls whether MoltenVK should treat a lost VkDevice as resumable, unless the
 	 * corresponding VkPhysicalDevice has also been lost. The VK_ERROR_DEVICE_LOST error has
 	 * a broad definitional range, and can mean anything from a GPU hiccup on the current
-	 * command buffer submission, to a phyically removed GPU. In the case where this error does
+	 * command buffer submission, to a physically removed GPU. In the case where this error does
 	 * not impact the VkPhysicalDevice, Vulkan requires that the app destroy and re-create a new
 	 * VkDevice. However, not all apps (including CTS) respect that requirement, leading to what
 	 * might be a transient command submission failure causing an unexpected catastophic app failure.
-	 * If this setting is enabled, in the case of a VK_ERROR_DEVICE_LOST error that does not
-	 * impact the VkPhysicalDevice, MoltenVK will remove the error condition on the VkDevice after
-	 * the current queue submission is finished, allowing the VkDevice to continue to be used.
-	 * If this setting is disabled, MoltenVK will maintain the VK_ERROR_DEVICE_LOST error condition
-	 * on the VkDevice, and subsequent use of that VkDevice will be reduced or prohibited.
 	 *
-	 * The value of this parameter should be changed before creating a VkDevice
-	 * that will use it, for the change to take effect.
+	 * If this setting is enabled, in the case of a VK_ERROR_DEVICE_LOST error that does NOT impact
+	 * the VkPhysicalDevice, MoltenVK will log the error, but will not mark the VkDevice as lost,
+	 * allowing the VkDevice to continue to be used. If this setting is disabled, MoltenVK will
+	 * mark the VkDevice as lost, and subsequent use of that VkDevice will be reduced or prohibited.
+	 *
+	 * The value of this parameter may be changed at any time during application runtime,
+	 * and the changed value will affect the error behavior of subsequent command submissions.
 	 *
 	 * The initial value or this parameter is set by the
 	 * MVK_CONFIG_RESUME_LOST_DEVICE
 	 * runtime environment variable or MoltenVK compile-time build setting.
-	 * If neither is set, this setting is disabled by default, and MoltenVK will not
-	 * resume a VkDevice that enters the VK_ERROR_DEVICE_LOST error state.
+	 * If neither is set, this setting is disabled by default, and MoltenVK
+	 * will mark the VkDevice as lost when a command submission failure occurs.
 	 */
 	VkBool32 resumeLostDevice;